Switch to a locked-by-default approach
