+++ app/controllers/schools/confirmed_bookings/upcoming_controller.rb
@@ -1,8 +1,6 @@
 module Schools
   module ConfirmedBookings
     class UpcomingController < ConfirmedBookingsController
-      include Schools::RestrictAccessUnlessOnboarded
-
       def index
         @bookings = current_school
           .bookings
