+++ app/controllers/schools/placement_requests/acceptance/add_more_details_controller.rb
@@ -2,7 +2,6 @@ module Schools
   module PlacementRequests
     module Acceptance
       class AddMoreDetailsController < Schools::BaseController
-        include Schools::RestrictAccessUnlessOnboarded
         before_action :set_placement_request
         before_action :ensure_previous_step_complete
