+++ app/controllers/schools/placement_requests/cancellations/notification_deliveries_controller.rb
@@ -2,7 +2,6 @@ module Schools
   module PlacementRequests
     module Cancellations
       class NotificationDeliveriesController < Schools::BaseController
-        include Schools::RestrictAccessUnlessOnboarded
         before_action :set_placement_request
         before_action :fetch_gitis_contact_for_placement_request
         before_action :ensure_placement_request_is_open, except: :show
