+++ app/controllers/schools/placement_requests/upcoming_controller.rb
@@ -1,8 +1,6 @@
 module Schools
   module PlacementRequests
     class UpcomingController < PlacementRequestsController
-      include Schools::RestrictAccessUnlessOnboarded
-
       def index
         @placement_requests = placement_requests.open
         assign_gitis_contacts @placement_requests
