+++ app/controllers/schools/placement_requests_controller.rb
@@ -1,7 +1,5 @@
 module Schools
   class PlacementRequestsController < Schools::BaseController
-    include Schools::RestrictAccessUnlessOnboarded
-
     def index
       @placement_requests = placement_requests
       assign_gitis_contacts(@placement_requests)
