+++ features/schools/availability_preferences/edit.feature
@@ -5,6 +5,7 @@ Feature: Editing placement dates
 
     Background:
         Given I am logged in as a DfE user
+        And my school is fully-onboarded
         And my school is set to use 'fixed' dates
 
     Scenario: Page title
