+++ features/schools/placement_dates/edit.feature
@@ -5,6 +5,7 @@ Feature: Editing placement dates
 
     Background:
         Given I am logged in as a DfE user
+        And my school is fully-onboarded
 
     Scenario: Page title
         Given I am on the edit page for my placement
