+++ features/schools/placement_dates/index.feature
@@ -5,6 +5,7 @@ Feature: Listing placement dates
 
     Background:
         Given I am logged in as a DfE user
+        And my school is fully-onboarded
 
     Scenario: Page title
         Given I am on the 'placement dates' page
