+++ features/schools/placement_dates/new.feature
@@ -5,6 +5,7 @@ Feature: Creating new placement dates
 
     Background:
         Given I am logged in as a DfE user
+        And my school is fully-onboarded
 
     Scenario: Page title
         Given I am on the 'new placement date' page
