+++ features/schools/toggle_enabling_and_disabling.feature
@@ -5,6 +5,7 @@ Feature: Toggling being enabled and disabled
 
     Background:
         Given I am logged in as a DfE user
+        And my school is fully-onboarded
 
     Scenario: Page title
         Given I am on the 'toggle requests' page
