+++ spec/controllers/schools/confirmed_bookings/cancellations_controller_spec.rb
@@ -4,7 +4,6 @@ require Rails.root.join("spec", "controllers", "schools", "session_context")
 describe Schools::ConfirmedBookings::CancellationsController, type: :request do
   include_context "logged in DfE user"
   include_context "stubbed out Gitis"
-  include_context "restricted unless school onboarded"
 
   let :school do
     Bookings::School.find_by!(urn: urn).tap do |s|
