+++ spec/controllers/schools/placement_requests/acceptance/confirm_booking_controller_spec.rb
@@ -4,7 +4,6 @@ require Rails.root.join('spec', 'controllers', 'schools', 'session_context')
 describe Schools::PlacementRequests::Acceptance::ConfirmBookingController, type: :request do
   include_context "logged in DfE user"
   include_context "stubbed out Gitis"
-  include_context "restricted unless school onboarded"
 
   let!(:pr) { create(:bookings_placement_request, school: @current_user_school) }
   before { create(:bookings_profile, school: @current_user_school) }
