--- spec/support/schools/restricted_unless_school_onboarded.rb
@@ -1,5 +1,0 @@
-shared_context 'restricted unless school onboarded' do
-  specify 'should include concern RestrictAccessUnlessOnboarded' do
-    expect(subject).is_a?(Schools::RestrictAccessUnlessOnboarded)
-  end
-end
